+++ src/main/java/jr_course/service/WordServiceImpl.java
@@ -1,13 +1,18 @@
 package jr_course.service;
 
+import com.fasterxml.jackson.databind.ObjectMapper;
 import jr_course.dao.WordRepository;
 import jr_course.entity.User;
 import jr_course.entity.Word;
+import org.json.JSONObject;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import jr_course.exception.*;
 
+import java.io.IOException;
+import java.io.StringReader;
 import java.util.List;
 import java.util.Optional;
 
@@ -33,10 +38,12 @@ public class WordServiceImpl implements WordService {
 	public List<Word> findAllByLevel(String level) {
 		logger.info("\"findByLevelContains(level)\"");
 		logger.info("Find all words in DB depending on the \"" + level + "\" level.");
+		if (!(level.equals("easy") || level.equals("medium") || level.equals("hard")))
+			throw new IncorrectDataInputException("Incorrect level data input - " + level);
+
 		return wordRepository.findAllByLevel(level);
 	}
 
-	// getOne?
 	@Override
 	public Word findById(int id) {
 		logger.info("\"findById(id)\"");
@@ -49,8 +56,8 @@ public class WordServiceImpl implements WordService {
 			word = result.get();
 		}
 		else {
-			logger.warn("Word not found.");
-			throw new RuntimeException("Слово с id - " + id + " не было найдено.");
+			logger.warn("Word was not found.");
+			throw new DataNotFoundException("Word with id " + id + " was not found.");
 		}
 
 		logger.info("Return word.");
@@ -58,17 +65,57 @@ public class WordServiceImpl implements WordService {
 	}
 
 	@Override
-	public void save(Word word) {
-		logger.info("\"save(word)\"");
-		logger.info("Save a word " + word + ".");
+	public Word save(String newWord) {
+		logger.info("\"save(newWord)\"");
+		logger.info("Save a word " + newWord + ".");
+
+		JSONObject root = new JSONObject(newWord);
+		String level = root.getString("level");
+		String jpKanji = root.getString("jpKanji");
+		String jpKana = root.getString("jpKana");
+		String ruWord = root.getString("ruWord");
+		String description = root.getString("description");
+
+		if (!(level.equals("easy") || level.equals("medium")
+				|| level.equals("hard")))
+			throw new IncorrectDataInputException("Incorrect level data input - " + level);
+
+		if (!jpKanji.isEmpty() && !jpKanji.matches("^[\\p{sc=Han}\\p{sc=Hiragana}\\p{sc=Katakana}]+$"))
+			throw new IncorrectDataInputException("Incorrect jpKanji data input - " + jpKanji);
+
+		if (!jpKana.matches("^[\\p{sc=Hiragana}\\p{sc=Katakana}]+$"))
+			throw new IncorrectDataInputException("Incorrect jpKana data input - " + jpKana);
+
+		if (!ruWord.matches("^([а-яА-Я0-9()/.,-]+(\\s)?)+$"))
+			throw new IncorrectDataInputException("Incorrect ruWord data input - " + ruWord);
+
+		if (description.trim().isEmpty())
+			throw new IncorrectDataInputException("Incorrect description data input - " + description);
+
+		ObjectMapper mapper = new ObjectMapper();
+		StringReader reader = new StringReader(newWord);
+
+		Word word = null;
+		try {
+			word = mapper.readValue(reader, Word.class);
+			logger.info("Word was read.");
+		} catch (IOException e) {
+			logger.debug(e.getMessage());
+			logger.debug("Incorrect data input - " + newWord);
+		}
+
 		wordRepository.save(word);
+
+		logger.info("Word was saved!");
+		return word;
 	}
 
 	@Override
 	public void deleteById(int id) {
 		logger.info("\"deleteById(id)\"");
-		logger.info("Delete a word in DB by id " + id + ".");
-		wordRepository.deleteById(id);
+		logger.info("Delete a word by id " + id + ".");
+		if(wordRepository.existsById(id)) wordRepository.deleteById(id);
+		else throw new DataNotFoundException("Word with id " + id + " was not found.");
 	}
 
 	@Override
@@ -84,12 +131,16 @@ public class WordServiceImpl implements WordService {
 
 		Word word = findById(wordId);
 
-		logger.info("Add word.");
-		word.addUser(user);
+		if (word.getUserCollection().contains(user)) {
+			logger.info("Word was already added!");
+		} else {
+			logger.info("Add word.");
+			word.addUser(user);
 
-		logger.info("Update word.");
-		wordRepository.save(word);
-		logger.info("Word was added to personal vocabulary.");
+			logger.info("Update word.");
+			wordRepository.save(word);
+			logger.info("Word was added to personal vocabulary.");
+		}
 	}
 
 	@Override
@@ -99,11 +150,14 @@ public class WordServiceImpl implements WordService {
 
 		logger.info("Find word by id and delete user.");
 		Word word = findById(wordId);
-		word.deleteUser(user);
 
-		logger.info("Update word.");
-		wordRepository.save(word);
-		logger.info("Word was deleted from personal vocabulary!");
+		if (word.getUserCollection().contains(user)) {
+			word.deleteUser(user);
+
+			logger.info("Update word.");
+			wordRepository.save(word);
+			logger.info("Word was deleted from personal vocabulary!");
+		} else logger.info("Word with id + " + wordId + " was not found in the personal list." );
 	}
 
 	@Override
